feat/ home club screen fixed
